pyproject.toml
[project]
name = "quattroformaggi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "maksnowak", email = "59841779+maksnowak@users.noreply.github.com" }
]
requires-python = ">=3.12.3"
dependencies = [
    "anthropic>=0.96.0",
    "claude-agent-sdk>=0.1.63",
    "pydantic>=2.10.6",
    "python-dotenv>=1.2.2",
]

[project.scripts]
quattroformaggi = "quattroformaggi.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
